+++ includes/acl/compression/impl/track_stream.h
@@ -103,7 +103,7 @@ namespace acl
 
 			TrackStream(iallocator& allocator, uint32_t num_samples, uint32_t sample_size, float sample_rate, animation_track_type8 type, track_format8 format, uint8_t bit_rate)
 				: m_allocator(&allocator)
-				, m_samples(reinterpret_cast<uint8_t*>(allocator.allocate(sample_size * num_samples + k_padding, 16)))
+				, m_samples(reinterpret_cast<uint8_t*>(allocator.allocate(sample_size * size_t(num_samples) + k_padding, 16)))
 				, m_num_samples(num_samples)
 				, m_sample_size(sample_size)
 				, m_sample_rate(sample_rate)
@@ -129,7 +129,7 @@ namespace acl
 			~TrackStream()
 			{
 				if (m_allocator != nullptr)
-					m_allocator->deallocate(m_samples, m_sample_size * m_num_samples + k_padding);
+					m_allocator->deallocate(m_samples, m_sample_size * size_t(m_num_samples) + k_padding);
 			}
 
 			TrackStream& operator=(const TrackStream&) = delete;
@@ -152,7 +152,7 @@ namespace acl
 				if (m_allocator != nullptr)
 				{
 					copy.m_allocator = m_allocator;
-					copy.m_samples = reinterpret_cast<uint8_t*>(m_allocator->allocate(m_sample_size * m_num_samples + k_padding, 16));
+					copy.m_samples = reinterpret_cast<uint8_t*>(m_allocator->allocate(m_sample_size * size_t(m_num_samples) + k_padding, 16));
 					copy.m_num_samples = m_num_samples;
 					copy.m_sample_size = m_sample_size;
 					copy.m_sample_rate = m_sample_rate;
